.agent_complex/requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Graph neural networks
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Chemistry
rdkit-pypi>=2022.9.1

# Mamba (state-space models)
mamba-ssm>=1.0.0
causal-conv1d>=1.0.0

# Transformers and datasets
transformers>=4.30.0
datasets>=2.12.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
tqdm>=4.65.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0

# API deployment
flask>=2.3.0
gunicorn>=20.1.0
